[metrics] Report metrics when marketplace operator-sources fail to reconcile
Problem:
The OpenShift Marketplace does not have a method to communicate to the
user, when an operator source cannot connect to quay.io due to
connectivity issues.

Solution:
This PR adds the custom metrics collection logic for three major operator
sources - redhat, community and certified. When an operator source fails
to reconcile, the respective registry-namespace and the operator-source
name is reported by the prometheus. These metrics are further whitelisted
to add them to telemeter, so as to create an alert.