[metrics] Report metrics for default operator-sources.
Problem:
The OpenShift Marketplace does not have a method to communicate to
the user, when an operator source cannot connect to quay.io due to
connectivity issues.

Solution:
This commit adds the custom metrics collection logic for three major
operator sources - redhat, community and certified. When an operator
source fails to reconcile, the respective error message and the
operator-source name is reported by the prometheus. Further, alerts
are configured for these metrics, so that they can be integrated
with the Telemetry server.